requirements.txt
streamlit==1.44.0
langchain-community==0.0.38
langchain-core==0.1.27
langchain-openai==0.0.8
langchain-text-splitters==0.0.2
openai==1.12.0
PyYAML==6.0.2
pymupdf==1.23.26
python-dateutil==2.9.0.post0 